move infochan node to ipfs_common #60
