+++ ipfs_common/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(robonomics_msgs)
+project(ipfs_common)
 
 ## Compile as C++11, supported in ROS Kinetic and newer
 # add_compile_options(-std=c++11)
@@ -10,12 +10,12 @@ project(robonomics_msgs)
 find_package(catkin REQUIRED
     COMPONENTS
         rospy
+        robonomics_msgs
         message_generation
 )
 
 ## System dependencies are found with CMake's conventions
 # find_package(Boost REQUIRED COMPONENTS system)
-
 
 ## Uncomment this if the package has a setup.py. This macro ensures
 ## modules and global scripts declared therein get installed
@@ -47,18 +47,13 @@ catkin_python_setup()
 ##   * add every package in MSG_DEP_SET to generate_messages(DEPENDENCIES ...)
 
 ## Generate messages in the 'msg' folder
-add_message_files(
-    FILES
-        Demand.msg
-        Offer.msg
-        Result.msg
-)
+# add_message_files(
+#   FILES
+# )
 
 ## Generate services in the 'srv' folder
 # add_service_files(
 #   FILES
-#   Service1.srv
-#   Service2.srv
 # )
 
 ## Generate actions in the 'action' folder
@@ -69,7 +64,10 @@ add_message_files(
 # )
 
 ## Generate added messages and services with any dependencies listed here
-generate_messages()
+generate_messages(
+    DEPENDENCIES
+        robonomics_msgs
+)
 
 ################################################
 ## Declare ROS dynamic reconfigure parameters ##
@@ -102,8 +100,8 @@ generate_messages()
 ## DEPENDS: system dependencies of this project that dependent projects also need
 catkin_package(
 #  INCLUDE_DIRS include
-#  LIBRARIES robonomics_msgs
-#  CATKIN_DEPENDS std_msgs
+#  LIBRARIES ros_erc20
+#  CATKIN_DEPENDS other_catkin_pkg
 #  DEPENDS system_lib
 )
 
@@ -115,13 +113,12 @@ catkin_package(
 ## Your package locations should be listed before other locations
 include_directories(
 # include
+# ${catkin_INCLUDE_DIRS}
 )
-
-catkin_package()
 
 ## Declare a C++ library
 # add_library(${PROJECT_NAME}
-#   src/${PROJECT_NAME}/robonomics_msgs.cpp
+#   src/${PROJECT_NAME}/ros_erc20.cpp
 # )
 
 ## Add cmake target dependencies of the library
@@ -132,7 +129,8 @@ catkin_package()
 ## Declare a C++ executable
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
-# add_executable(${PROJECT_NAME}_node src/robonomics_msgs_node.cpp)
+# add_executable(${PROJECT_NAME}_node src/ros_erc20_node.cpp)
+# add_executable(${PROJECT_NAME}_node src/ros_eth_node.cpp)
 
 ## Rename C++ executable without prefix
 ## The above recommended prefix causes long target names, the following renames the
@@ -184,12 +182,28 @@ catkin_package()
 #   DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}
 # )
 
+catkin_package()
+
+catkin_install_python(
+    PROGRAMS
+        scripts/ipfs_channel_node
+    DESTINATION
+        ${CATKIN_PACKAGE_BIN_DESTINATION}
+)
+
+install(
+    FILES
+        launch/ipfs_channel.launch
+    DESTINATION
+        ${CATKIN_PACKAGE_SHARE_DESTINATION}/launch
+)
+
 #############
 ## Testing ##
 #############
 
 ## Add gtest based cpp test target and link libraries
-# catkin_add_gtest(${PROJECT_NAME}-test test/test_robonomics_msgs.cpp)
+# catkin_add_gtest(${PROJECT_NAME}-test test/test_ros_erc20.cpp)
 # if(TARGET ${PROJECT_NAME}-test)
 #   target_link_libraries(${PROJECT_NAME}-test ${PROJECT_NAME})
 # endif()
