+++ ipfs_common/package.xml
@@ -1,24 +1,26 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>robonomics_msgs</name>
+  <name>ipfs_common</name>
   <version>0.0.1</version>
-  <description>Robonomics core messages</description>
+  <description>Commonly used IPFS integration node</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
-  <maintainer email="akru@aira.life">Alexander Krupenkin</maintainer>
+  <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
+  <maintainer email="mail@akru.me">Alexander Krupenkin</maintainer>
+  <maintainer email="eilyichev@aira.life">Evgeny Ilyicev</maintainer>
 
 
   <!-- One license tag required, multiple allowed, one license per tag -->
   <!-- Commonly used license strings: -->
   <!--   BSD, MIT, Boost Software License, GPLv2, GPLv3, LGPLv2.1, LGPLv3 -->
-  <license>BSD</license>
+  <license>BSD3</license>
 
 
   <!-- Url tags are optional, but multiple are allowed, one per tag -->
   <!-- Optional attribute type can be: website, bugtracker, or repository -->
   <!-- Example: -->
-  <!-- <url type="website">http://wiki.ros.org/robonomics_msgs</url> -->
+  <!-- <url type="website">http://wiki.ros.org/ros_erc20</url> -->
 
 
   <!-- Author tags are optional, multiple are allowed, one per tag -->
@@ -35,24 +37,19 @@
   <!--   Note that this is equivalent to the following: -->
   <!--   <build_depend>roscpp</build_depend> -->
   <!--   <exec_depend>roscpp</exec_depend> -->
-  <!-- Use build_depend for packages you need at compile time: -->
-  <!--   <build_depend>message_generation</build_depend> -->
   <!-- Use build_export_depend for packages you need in order to build against this package: -->
   <!--   <build_export_depend>message_generation</build_export_depend> -->
   <!-- Use buildtool_depend for build tool packages: -->
   <!--   <buildtool_depend>catkin</buildtool_depend> -->
-  <!-- Use exec_depend for packages you need at runtime: -->
-  <!--   <exec_depend>message_runtime</exec_depend> -->
   <!-- Use test_depend for packages you need only for testing: -->
   <!--   <test_depend>gtest</test_depend> -->
   <!-- Use doc_depend for packages you need only for building documentation: -->
   <!--   <doc_depend>doxygen</doc_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>message_generation</build_depend>
+  <build_depend>robonomics_msgs</build_depend>
 
-  <!-- The export tag contains other, unspecified, tags -->
-  <export>
-    <!-- Other tools can request additional information be placed here -->
+  <exec_depend>message_runtime</exec_depend>
+  <exec_depend>robonomics_msgs</exec_depend>
 
-  </export>
 </package>
