+++ ipfs_common/scripts/ipfs_channel_node
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from robonomics_lighthouse.nodes import infochan_node
+from ipfs_common.nodes import ipfs_channel_node
 
 if __name__ == '__main__':
-    infochan_node()
+    ipfs_channel_node()
