+++ ipfs_common/src/ipfs_common/nodes.py
@@ -1,0 +1,4 @@
+from . import ipfs_channel
+
+def ipfs_channel_node():
+    ipfs_channel.IPFSChannel().spin()
