+++ robonomics_liability/tests/liability.test
@@ -17,16 +17,20 @@
 
   <!-- liability group -->
   <group ns="liability">
-    <!-- message signer interface -->
-    <include ns="infochan" file="$(find ethereum_common)/launch/signer.launch">
-      <arg name="keyfile" value="$(arg keyfile)" />
-      <arg name="keyfile_password_file" value="$(arg keyfile_password_file)" />
-    </include>
 
-    <include file="$(find robonomics_lighthouse)/launch/infochan.launch">
-      <arg name="lighthouse_contract" value="$(arg lighthouse_contract)" />
-      <arg name="ipfs_http_provider" value="$(arg ipfs_http_provider)" />
-    </include>
+    <!-- Information channel interface -->
+    <group ns="infochan">
+      <include  file="$(find ipfs_common)/launch/ipfs_channel.launch">
+        <arg name="lighthouse_contract" value="$(arg lighthouse_contract)" />
+        <arg name="ipfs_http_provider" value="$(arg ipfs_http_provider)" />
+      </include>
+
+      <!-- message signer interface -->
+      <include file="$(find ethereum_common)/launch/signer.launch">
+        <arg name="keyfile" value="$(arg keyfile)" />
+        <arg name="keyfile_password_file" value="$(arg keyfile_password_file)" />
+      </include>
+    </group>
 
     <node pkg="robonomics_liability" type="listener_node" name="listener" output="screen">
       <param name="liability_contract_abi" textfile="$(find robonomics_liability)/abi/liability.json" />
