--- robonomics_lighthouse/launch/infochan.launch
@@ -1,16 +1,0 @@
-<launch>
-
-  <arg name="lighthouse_contract" default="airalab.lighthouse.3.robonomics.eth" />
-  <arg name="ipfs_http_provider" default="http://127.0.0.1:5001" />
-
-  <!-- Robonomics information channel -->
-  <group ns="infochan">
-
-    <!-- PubSub information channel interface -->
-    <node pkg="robonomics_lighthouse" type="infochan_node" name="channel" output="screen">
-      <param name="ipfs_http_provider" value="$(arg ipfs_http_provider)" />
-      <param name="lighthouse_contract" value="$(arg lighthouse_contract)" />
-    </node>
-
-  </group>
-</launch>
