+++ robonomics_lighthouse/launch/lighthouse.launch
@@ -15,16 +15,18 @@
   <group ns="lighthouse">
 
     <!-- Information channel interface -->
-    <include file="$(find robonomics_lighthouse)/launch/infochan.launch">
-      <arg name="lighthouse_contract" value="$(arg lighthouse_contract)" />
-      <arg name="ipfs_http_provider" value="$(arg ipfs_http_provider)" />
-    </include>
+    <group ns="infochan">
+      <include  file="$(find ipfs_common)/launch/ipfs_channel.launch">
+        <arg name="lighthouse_contract" value="$(arg lighthouse_contract)" />
+        <arg name="ipfs_http_provider" value="$(arg ipfs_http_provider)" />
+      </include>
 
-    <!-- message signer interface -->
-    <include ns="infochan" file="$(find ethereum_common)/launch/signer.launch">
-      <arg name="keyfile" value="$(arg keyfile)" />
-      <arg name="keyfile_password_file" value="$(arg keyfile_password_file)" />
-    </include>
+      <!-- message signer interface -->
+      <include file="$(find ethereum_common)/launch/signer.launch">
+        <arg name="keyfile" value="$(arg keyfile)" />
+        <arg name="keyfile_password_file" value="$(arg keyfile_password_file)" />
+      </include>
+    </group>
 
     <!-- XRT token interface -->
     <group ns="xrt">
