+++ robonomics_msgs/setup.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+## ! DO NOT MANUALLY INVOKE THIS setup.py, USE CATKIN INSTEAD
+
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+# fetch values from package.xml
+setup_args = generate_distutils_setup(
+    packages=['robonomics_msgs'],
+    package_dir={'': 'src'})
+setup(**setup_args)
